CI: rm Python 3.3 from test environments
because `matplotlib == 2.0.2` requires Python 3.4 or later
in the 3.x series [1] (though it includes the Trove classifier
for Python 3.3 [2]).

The tests on Travis CI did not reveal this requirement of
`matplotlib` earlier because Python 3.6 was present within Miniconda,
thus the Python 3.3 environment on Travis CI installed the Python 3.6
wheel `matplotlib-2.0.2-cp36-cp36m-manylinux1_x86_64.whl`.

Besides, note that `pip install matplotlib` occurs under `script:`
within `.travis.yml`. As a consequence, failure to install
`matplotlib` is treated as a failed test, so the test run doesn't
stop as soon as the installation fails, but instead the
tests that depend on `matplotlib` fail to import it.

In addition, the planned EOL of Python 3.3 is Sep 29, 2017 [3].

[1] https://matplotlib.org/users/installing.html#required-dependencies
[2] https://github.com/matplotlib/matplotlib/issues/9085
[3] https://www.python.org/dev/peps/pep-0398/#x-end-of-life