Handle in readdstar.py absence of raw_input in Py3
readdstar.py, a script that loads the output of LTL2DSTAR into a
NetworkX DiGraph, can read input from stdin. This changeset makes it
robust to the superseding of `raw_input` by `input` in Python 3.

This code is in the contrib/ directory, so it was not included in the
general upgrade of TuLiP to support Python 3. (pull request #180)

[ci skip]